+++ packages/graph-explorer/src/connector/gremlin/templates/keywordSearchTemplate.ts
@@ -9,6 +9,7 @@ import type { KeywordSearchRequest } from "../../AbstractConnector";
  * searchByAttributes = ["city", "code"]
  * limit = 100
  * offset = 0
+ * exactMatch = false
  *
  * g.V()
  *  .hasLabel("airport")
@@ -25,6 +26,7 @@ const keywordSearchTemplate = ({
   searchByAttributes = [],
   limit = 10,
   offset = 0,
+  exactMatch = false,
 }: KeywordSearchRequest): string => {
   let template = "g.V()";
 
@@ -43,7 +45,13 @@ const keywordSearchTemplate = ({
       .filter(attr => attr !== "__all")
       .map(attr => {
         if (attr === "_id") {
+          if (exactMatch === true) {
+             return `has(id,"${searchTerm}")`;
+          }
           return `has(id,containing("${searchTerm}"))`;
+        }
+        if (exactMatch === true) {
+          return `has("${attr}","${searchTerm}")`;
         }
         return `has("${attr}",containing("${searchTerm}"))`;
       })